requirements.txt
FLask
werkzeug
openpyxl
python-docx
python-pptx
PyPDF2
web-browser
python-dotenv
fpdf